requirements.txt
bokeh
scikit-learn
pandas
scipy
tensorflow-cpu